requirements.txt
flask
flask-restx
flask-sqlalchemy
mysqlclient
